Catch Python exceptions and report them along with their tracebacks.
Currently, when a Python method called by the C++ code raises an
exception, say for instance, resourceOffer method in the Scheduler
class raises an exception, SWIG only complains that there was some
exception it could notice and reports all such exceptions as
Swig::DirectorMethodException. An issue for this has been reported
in issue #44. This patch adds the SWIG interface required to catch
all the Python exceptions and report them along with the resulting
traceback and the exception string(or the exception value).

Closes gh-44

git-svn-id: https://svn.apache.org/repos/asf/incubator/mesos/trunk@1132238 13f79535-47bb-0310-9956-ffa450edef68