Be more intelligent about slot allocation.
When we get a resource offer, we'll try to be clever about how we
allocate slots to the offers.  Ideally we want to use as much of the
resources available when we have a lot of pending tasks, and we also
don't want to use more resources than we actually need.

Review: https://reviews.apache.org/r/13078