Allowed specification of mininum slot count.
To ensure Hadoop jobs begin promptly, we can specify a minimum number of
'hot slots' to be available for use.  This addresses the TaskTracker
spin up delay that exists with Hadoop on Mesos.  This can be a nuisance
with lower latency applications, such as ad-hoc Hive queries.

Review: https://reviews.apache.org/r/12921