Fixed the CgroupsIsolator to check for the existence of the nested "test" cgroup.
Sometimes the cgroup tests fail because the "test" cgroup is already
there. This ensures that it gets cleaned by the next test.

We also check that it exists before we try creating it.  If it
already exists, we skip the creation.

From: Brenden Matthews <brenden@diddyinc.com>
Review: https://reviews.apache.org/r/11126