Added link to executor sandbox from framework template
The framework page is rendered with info from the master, but the
sandbox directory is known only by the slave. To link to the directory
from a page rendered by the master, create a page in between that acts
as a redirector by first asking for the directory from the slave.

If the sandbox is important for linking, the master could return
`executor.directory` in the `state.json` endpoint and negate the need
for this redirection stuff.

Review: http://reviews.apache.org/r/13006