Fixed the Balloon Framework Test.
Stop exporting variables that begin with MESOS_ into the mesos
environment. The mesos-master and mesos-slave binaries will not start
if there are unrecognized variables that begin with MESOS_ in the
environment.

Set the LD_LIBRARY_PATH so that the balloon binaries built with
libtool will be able to find all of their pieces and execute properly
without needing to be installed.

Cleanup the cleanup code so we don't unmount a cgroup hierarchy unless
we have fully cleaned it up. This is necessary to prevent a problem
where the hierarchy can never be mounted again.

Update the command line options to mesos-master to pass the hierarchy
and the mesos root directory in that hierarchy.

Clearly report when the balloon hierarchy returns with an unexpected
return code in order to indicate whether the test failed because of a
problem in the baloon framework or within mesos.

From: Eric Biederman <ebiederm@xmission.com>
Review: https://reviews.apache.org/r/13034