Preserved newlines for copy/paste in Firefox
Due to a difference in implementation, Firefox does not preserve
newlines on copy when output is styled with "white-space: pre-wrap;"[1].
Firefox does preserve newlines if the output is wrapped in <pre><code>.
Switching to the semantic markup fixes Firefox and functions the same
as before in Chrome and in Safari.

This change addresses MESOS-661.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=116083

Review: http://reviews.apache.org/r/13946