Fixed build breakage from 37e946683f1bb6a.
In 37e946683f1bb6a I fixed a build failure from a warning, which are
treated as errors.

Newer versions of GCC have a feature (documented here:
http://gcc.gnu.org/onlinedocs/gcc/Warning-Options.html) where unknown
warning flags are ignored.

In older version of GCC (such as 4.1 or 4.2), it appears that these flags
aren't ignored, resulting in an error.

Review: https://reviews.apache.org/r/14164