Task reconciliation for frameworks.
A framework can reconcile tasks in connection with a master fail-over
by sending it's "last known state" of its tasks and confirm or
reestablish the current state of the setup.

If existing tasks have changed state, a status update with the new
state will be sent.

Soon, we may be able to reliably tell if any slave or task is no
longer present. A status update with TASK_LOST should be sent to the
framework.

From: Niklas Nielsen <nik@qni.dk>
Review: https://reviews.apache.org/r/14540