Removed unimplemented Dashboard page.
The Dashboard page has been stubbed in the webUI for a while but has not
received any attention yet. It is misleading when a user clicks on the
navigation and doesn't understand what the graphs mean.

The files are not used for anything else, and so they can safely be
removed without affecting the rest of the UI. D3 and Cubism can be
re-added later if they prove useful.

Review: https://reviews.apache.org/r/14657