Added a C++ style checker.
To avoid style glitches, it would be handy to have a style checker to
catch the most common (and hard to find) errors such as whitespaces,
tabs, line widths and so forth.

While it is not likely to find a fully compatible style checker,
Google's cpplint can be configured to match a small subset of rules.

This patch includes a wrapper, support/mesos-style.py, which runs
cpplint on mesos source (excluding 3rdparty libraries such as glog,
protobuf, boost and so on).

For now, only the whitespace/tab rule is active. When style errors
gets fixed, we can increase the verboseness of the applied rules.

From: Niklas Nielsen <nik@qni.dk>
Review: https://reviews.apache.org/r/14559