Added a flash service for UI alerts after redirects.
When an error occurs during a redirect, it is useful to show an error
message on the following page view rather than the current one. In
particular, when clicking a sandbox link for a slave that is
inaccessible, the `SlaveExecutorRerouterCtrl` sends a request to the
desired slave and redirects back to root if an error occurs. It's not
possible to show an error during the redirect, and so rendering on
the root view is ideal.

Flash alerts are rendered once and then cleaned up; they will not
display more than once.

Review: http://reviews.apache.org/r/15198