Fixed a bug where Group::cancel can return a failed future if the membership is already cancelled.
The API stipulates that a successful Future<bool> but with 'false'
value should be returned.

This can happen when:
1) The cancel() operation is queued up due to connectivity issue but
then invoked again by sync() the membership could be gone.
2) Even if not delayed, in cancel() we first checked whether the
membership exists in our local cache but the membership could have
expired by then (or when we invoke zk->remove(...)) but we have yet to
receive the update for it.

From: Jiang Yan Xu <yan@jxu.me>
Review: https://reviews.apache.org/r/15487