Made pailer scrollable in Chrome and Safari.
After a change to fix the pailer in Firefox, Chrome and Safari broke.
By removing the inner <code> block and using just a <pre> element, the
calculations done by the pailer work again and in all 3 browsers.

The style changes simplified the CSS.

In JS, references to the DOM are kept in variables to prevent unneeded
trips to the DOM on window resizing.

Review: http://reviews.apache.org/r/16029