Moved control of navbar tabs to MainController
This removes another dependency on jQuery and lets the MainController
control its tabs. `setNavbarActiveTab` was a global function that was
accessing the global DOM, which meant the controllers were editing
things they didn't control.

The MainController now sets the tab based on the path.

We are almost to zero jQuery...

Review: http://reviews.apache.org/r/16395