Add post function to http.hpp of libprocess.
This adds the post equivalent of the get function to http.hpp.

The existing get method is refactored into internal::request with
arguments for method, path, query, contentType, and body.

The get and post functions can then be implemented as simple wrappers
of internal::request.

The new post function has all required arguments of path, contentType,
and body.

The get function's query argument is also changed to be Option<string>
to better indicate that it is optional.

That change requires adding a forwarding overload with the old
signature (but w/o the default parameter value) for back-compat.

There are also new unit tests to verify the existing behavior of get
and the new post behavior.

From: Charlie Carson <ccarson@twitter.com>
Review: https://reviews.apache.org/r/16839