Added launchTasks for list of offers
Running tasks on more than one offer belonging to a single slave can
be useful in situations with multiple out-standing offers.

This patch extends the usual launchTasks() to accept a vector of
OfferIDs. The previous launchTasks (accepting a single OfferID) has
been kept for backward compatibility, but this now calls the new
launchTasks() with a one-element list. This also applied for the JNI
and python interfaces, which accepts both formats as well.

Offers are verified to belong to the same slave and framework,
before resources are merged and used.

Review: https://reviews.apache.org/r/14669