Introduced a Repairer class to master.
Add a Repairer class to master.  Currently it is just a stub, but
it will soon get wired up to the new observe end-point in master and
recieve slave state updates.  From there, it will be responsible for
persisting the state and ultimately performing repairs on unhealthy
slaves.

The Repairer has virtual methods and destructor so that it can
easily be mocked for testing.

Updated the existing master constructor to take a Repairer* and
local/local.cpp, master/main.cpp, & tests/cluster.hpp to new up a
Repairer, pass it to master's constructor and delete the Repairer when
it's finished.

Review: https://reviews.apache.org/r/17564