*** MODIFIED for 0.18.1-rc2 *** Add a new boolean CommandInfo.URI.extract.
This is a patch that includes a new `extract` boolean value on
CommandInfo.URI to toggle the auto-extraction of compressed archives
when shipping URIs with an executor or task.

It's worth noting that this changes the behaviour of the
MESOS_EXECUTOR_URI environment variable used. I've added some comments
which outlines the expected format, instead of (e.g)
hdfs://foobar/testing.tar.gz+0 (1 signifying it's not executable)
there is now an extra flag on the end (X or N) defining whether to
auto-extract or not, resulting in hdfs://foobar/testing.tar.gz+0X by
default.

Review: https://reviews.apache.org/r/20674