Updated policy for building against system dependencies.
Allows preventing the use of bundled libraries in favor of
preinstalled versions, where possible.

Using --disable-bundled generally switches from bundled to
preinstalled libraries.

If --disable-bundled is supplied or --with-leveldb[=DIR] was given but
the configuration phase could not locate the dependency, it stops with
an error.

If --disable-bundled is not supplied and no --with-leveldb[=DIR] was
given, the bundled library is built and used (DEFAULT).

NOTE:

1. This review only implements this preinstalled library usage on
LevelDB & ZooKeeper. Subsequent review requests for other bundled
libraries will follow as soon as this implementation has been
accepted.

2. This modification removed the old --without-installed-PACKAGE
semantics, as they conflict with well established --with-PACKAGE
semantics

Review: https://reviews.apache.org/r/20119