Fixed ExternalContainerizer Subprocess use for non sandboxed invocations.
When using Subprocess on non sandboxed invocations of the external
containerizer program, stderr was not getting redirected at all,
possibly triggering deadlocks due to a filled stderr pipe that was
not getting drained.
This patch redirects to /dev/null when no sandbox is given for an
invocation. See MESOS-1336.

Review: https://reviews.apache.org/r/21675