Added healthy field to TaskUpdate.
As a part of the health check story, this patch adds an optional
healthy field to the task update proto.  The semanticist should be
understood as, if the field is absent - no health check has (yet) been
carried out.  If present, it will either encode unhealthy
(healthy=false) or healthy determined by the HealthCheck protobuf in
the command info proto.  This means that multiple TASK_RUNNING updates
can be sent for a task, updating the state of the last health check.

Review: https://reviews.apache.org/r/21499