Improved failure handling of DockerContainerizer.
Two improvements:

(1) Checked the exit status of 'docker.run()' in order to fail early.

(2) Improved the override on mesos-executor to return the exit status
of the Docker container rather than the exit status of doing 'docker
wait' (which will be 0 even if the container exited with -1 because
'docker wait' correctly "waited" and it prints the container's exit
status to stdout).