Updated allocator to offer cpu only or memory only resources.
As already explained in JIRA MESOS-1688, there are schedulers
allocating memory only for the executor and not for tasks. For tasks
only CPU resources are allocated in this case. Such a scheduler does
not get offered any idle CPUs if the slave has nearly used up all
memory. This can easily lead to a dead lock (in the application, not
in Mesos).

Simple example:
1. Scheduler allocates all memory of a slave for an executor
2. Scheduler launches a task for this executor (allocating 1 CPU)
3. Task finishes: 1 CPU , 0 MB memory allocatable.
4. No offers are made, as no memory is left. Scheduler will wait for
offers forever. Dead lock in the application.

To fix this problem, offers must be made if CPU resources are
allocatable without considering allocatable memory.

Review: https://reviews.apache.org/r/25035