Install Python libraries during make install.
Because eggs are impossible to uninstall, pip + wheel has been added
as a packaging format. The wheels are generated in addition to eggs
and then subsequently installed/uninstalled.

Note that the building of eggs has not been removed. This is because
wheels cannot be run in place like eggs are via. modification of
PYTHONPATH. The eggs are needed for all the test scripts to run
correctly without actually installing anything locally.

Review: https://reviews.apache.org/r/24264