Add --with-curl to libprocess since <stout/net.hpp> needs it.
libprocess uses `<stout/net.hpp>` which includes `<curl/curl.h>`. On a machine without system-level libcurl installed, we should be able to build `mesos` with the `--with-curl=<path-to-libcurl>` flag. The `-I<path-to-libcurl>` gets added to the `CPPFLAGS` for `mesos` but it doesn't added for libprocess. This patch makes libprocess actually use the `--with-curl` flag.

Review: https://reviews.apache.org/r/26352