stout: Always use stout ABORT() rather than system abort()
This makes it so any time there is an abort, we get a line number and at least
a basic message as to why there was an abort. If you want a clean(er) exit, use
<stout/exit>.

Also adds an overload which takes a standard string and unwraps it to a
const char * automatically, since a lot of the time we are building strings
to pass them to abort.

Review: https://reviews.apache.org/r/26472