Created a Result(Try&) constructor for Result<T>.
Try<T> doesn't have a default constructor and so it's not possible to
have a Try<T> class member variable  unless we initialize it with
Error().

With this patch, we can instead have a Result<T> class member variable
which is initialized to None() and can be assigned a Try<T> type.

Review: https://reviews.apache.org/r/26955