Added a libprocess network benchmark.
The benchmark forks and launches multiple "libprocess" processes in
each child. They are aware of the "master"'s (parent) address and all
play ping pong with it. This benchmark measures throughput in terms of
the number of RPCs handled per second using persistent (linked)
connections.

A new test file (benchmarks) is introduced because we want to fork
before libprocess is initialized. This allows us to prevent
short-circuiting of message passing between processes under the same
ProcessManager. This way we force the execution path of the underlying
event management system.

Review: https://reviews.apache.org/r/26150