Fix OsTest.killtreeNoRoot check for reparenting.
Reparenting does not always assign to pid 1 (/sbin/init). If there is
a user init such as init --user with some other pid, this will be the
new parent.

Modify os_tests to check that the subtree has been reparented to a
process different from its original parent (a.k.a. child) and that it
is not a zombie.

Review: https://reviews.apache.org/r/27461