libprocess:  Replaced the ip and port pairs from UPID class and process namespace with Node class.
At the moment, the Node class is used to keep a mapping from a socket to the ip & port pair in the process namespace.
I want to propose to extend its use by replacing the ip & port fields from the UPID class and process namespace with this type.

Review: https://reviews.apache.org/r/27446