Made ProcessBase::route hard-fail rather than soft fail.
It is an API expectation that the path start with '/', and if it doesn't
it should be a programmer error.

No code actually checked the return code. If a developer did get this
wrong previously it would silently do nothing and in debugging their
endpoint just wouldn't appear. This makes it simple/easy for them to
find.

Review: https://reviews.apache.org/r/27660