Introduced Mesos Hooks abstraction.
Hooks are implemented as Mesos modules. The first hook is a task
label-decorator that is called during the task launch sequence in
Mesos master.

This hook allows hook modules to add additional labels to the incoming
TaskInfo object. The labels are then read on the slave/executor side
which may then act upon them.

Review: https://reviews.apache.org/r/28655