Out of tree build 4: Introduced ExecutorRunState struct for Isolators and Launchers.
The struct is constructed from some fields of RunState and passed on
to Isolators. This further allows us to do cleanup various
isolators/launchers. The goal was to remove the dependency on RunState
so that we can eventually expose slave/containerizer/isolator.hpp as a
public header.

Review: https://reviews.apache.org/r/30338