Reenabled hadoop_home and frameworks_home slave flags for mesos-fetcher.
Now the containerizer/fetcher sets the HADOOP_HOME variable for
mesos-fetcher again if the slave flag hadoop_home is set. Added a test
that checks that HDFS fetching is not broken and also ensures that the
flag gets translated to the environment variable and then gets applied
in mesos-fetcher. Created a mock hadoop implementation script for this.
This script has the exact same side effects as a real haddop client in
the scope of our testing. Using this, Mesos testing has no extra
external dependencies (on Hadoop).

Slave flag frameworks_home does not need to be an evironment variable.
It is now part of FetcherInfo only, but it also gets tested now that it
works.

Review: https://reviews.apache.org/r/31362