Fixed race in getFieldID helper.
The getFieldID helper looks up the java/lang/NoSuchFieldError class and
stores it in a static. It has turned out to provoke a racy behavior with
Java 8 when multiple drivers are created (and the class object may have
been created by another thread). This patch reverts the 'static'
optimization and looks up the class object when exceptions are thrown.

Review: https://reviews.apache.org/r/31818