Used LDADD to add unbundled libraries to all command line programs.
When building with unbundled libraries, every program needs to link
against the libraries, and can't assume it will be linked against
because the program links with a program which links against the
libraries.

This comes up on Debian and Ubuntu[1] which set the flag
`-Wl,--no-copy-dt-needed-entries` which forcibly disables the binaries
from getting automatically the dependencies of the shared libraries they
link against.

Also adds -Wl,--as-needed to AM_LDFLAGS to silence warnings in Fedora
packaging since we link against more libraries than things actually use
at the command line. Debian and Ubuntu default to this option.

OS X doesn't have `-Wl,--as-needed` or `--no-copy-dt-needed-entries` so
only enable it on Linux.

[1] https://wiki.ubuntu.com/ToolChain/CompilerFlags#A-Wl.2C--no-copy-dt-needed-entries

Review: https://reviews.apache.org/r/32008