Remove the checkpoint variable entirely from slave/flags.hpp.
As a number of tests rely on the checkpointing flag to be false, a few
tests had to be adapted. Removed the following test as the tested logic
is specific to (old) non-checkpointing slaves:
SlaveRecoveryTest.NonCheckpointingSlave: This test checks whether a
non-checkpointing slave is not scheduled to a checkpointing framework.
It can be removed as all slaves are now checkpointing slaves.

Review: https://reviews.apache.org/r/31539