Fixed sandbox ownership bug for executors without URIs.
During recent refactorings, executor directory ownership was delegated
to the fetcher. However, the fetcher is not invoked if no URIs are
present in the executor or task command. This left some of these tasks
broken as the directory ownership defaulted to the mesos-slave's (root).

Review: https://reviews.apache.org/r/32911