libprocess: Placed noreturn attribute correctly for C11.
Places bundled libev noreturn attribute correctly for C11.

The bundled libev uses different noreturn attributes based on whether
it is compiled with C11 or not. The C11 codepath places the noreturn
attribute in an incorrect (but accepted by GCC) location. Move it to
the standard location which is accepted by all supported compilers.

Review: https://reviews.apache.org/r/32747