Refactored common functionality into FlagsBase.
Jira: MESOS-2711

Every program that uses stout's `FlagsBase` ends up re-implementing
the `usage()` function, and adding a `bool help` (and associated
--help flag); this functionality has now been refactored in the base
class and is available everywhere.

This change attempts to be backward-compatible, so it does not alter
the behavior of the program when --help is invoked (by, e.g.,
automatically printing usage and exiting) but leaves up to the caller
to check for `flags.help` and then decide what action to take.

There is now a default behavior for the "leader" ("Usage: <program
name> [options]") but the client API allows to modify that too.

Note - anywhere I found the use of the `--help` flag the behavior was
the same: print usage and exit (see also
https://reviews.apache.org/r/34195).

Review: https://reviews.apache.org/r/34193