Refactoring to use FlagsBase common functionality.
Jira: MESOS-2711

All the main() methods have been refactored to use the definition of
FlagsBase::help flag and FlagsBase::usage().

This CL also tries to bring some uniformity to the use of exit codes:
if this is deemed to be worth making it uniform, we can come up with
common rules and extend the changes here to be compliant.

This touches a lot of files, but keep scrolling, and you will see a
pattern emerge.

Review: https://reviews.apache.org/r/34195