Added `version` string to MasterInfo.
Jira: MESOS-2957

Adds an optional version string to the `MasterInfo` protobuf,
to simplify handling versioning in HTTP API calls.

The version string is taken from <mesos/version.hpp> which is the
same used when starting up Master (master/main.cpp).

I've added a simple test on the back of the `MasterDetector` test,
as this was the place where we would expect the `MasterInfo` to
have been fully populated by real production code (as opposed to
other places, where it is actually handled by mocks).

Review: https://reviews.apache.org/r/36036