Slave exits gracefully on DNS lookup failure.
Jira: MESOS-2962

Slave fails with Abort stacktrace when DNS cannot resolve hostname.

If the DNS cannot resolve the hostname for a slave node, we correctly
return an Error object, but we then fail with a segfault.

This code adds a more user-friendly message and exits normally (with
an `EXIT_FAILURE` code).

For example, forcing `net::getIp()` to always return an Error, now
causes the slave to exit like this:

```
$ ./bin/mesos-slave.sh --master=10.10.1.121:5405
Failed to obtain the IP address for 'stratos.local'; the DNS service may not be able to resolve it: Name or service not known

$ echo $?
1
```

Review: https://reviews.apache.org/r/36061