Added support for cache control in libprocess when dealing with static files.
When serving a static file, libprocess returns the header
`Last-Modified` which is used by browsers to control Cache. When a http
request arrives containing the header `If-Modified-Since`, a response
`304 Not Modified` is returned if the date in the request and the
modification time (as returned by doing `stat` in the file) coincide.

Unit tests added.

Review: https://reviews.apache.org/r/30032