Add CMake-based build system for Mesos project.
See MESOS-898.

This is the first of two commits that will introduce a CMake-based
build system for the Mesos project. These commits are split along the
libprocess boundary -- this first commit will provide the CMake build
logic for the "core Mesos" project, while the second one will
introduce the CMake build logic for the process library. They are
committed independently to provide a natural history boundary between
the two projects.

CMake provides a number of advantages over the Make-based build system
currently in place. The most obvious of these is much greater platform
independence -- CMake supports a wide variety of platforms and
environments (even Visual Studio on Windows), which opens Mesos up for
contributions from organizations with drastically different toolchains
than the current Mesos default.

The second is that this gives us an opportunity to audit the existing
build system.

The plan moving forward is to develop the CMake build system
incrementally and in paralllel to the autoconf build system.

Review: https://reviews.apache.org/r/36513