Add CMake-based build system for the process library.
See MESOS-898.

This commit is the second in a series of two commits that will introduce
a CMake-based build system for Mesos. The first introduced the build
system for the "core Mesos" project; this one will introduce a similar
system for the process library.

For more details see the "core Mesos" commit.

Review: https://reviews.apache.org/r/36514