Introduced 'recordio' encoding facilities to stout.
Note that most "Record-IO" encodings are used for file I/O
and consequently use a fixed-size header to encode the record
length. However, decoding a base-10 integer is more
straightforward to implement in most languages, and so this
was chosen instead. (Note that the Twitter streaming API
uses the same technique for portability).

Review: https://reviews.apache.org/r/36677