Add CMake logic for building the Stout tests.
This project is the first of two commits that will add support for
building and running the Stout tests, using CMake. NOTE: This commit
will be "wired up" into the Mesos project in the second commit. It is
so split because of our strict requirements that the Stout source
history remain separated from the rest of the Mesos project history,
and it is not allowed to commit both these things at the same time.

More particularly, this commit will introduce the core CMake logic for
actually building the Stout tests from source, as well as the logic
for finding a couple dependencies we expect to be on the system
already (particularly SVN and APR).

Review: https://reviews.apache.org/r/36689