Enabling IP Discovery command.
JIRA: MESOS-2902

It is sometimes useful to enable an external script to configure the
IP address the Mesos Master will bind to on the server, where it's not
desirable to set the --ip flag and/or a "wrapper" script is not a
viable option.

This patch adds an --ip_discovery_command to point to a local script
that will emit as its only output the IP address that the Master will
bind to: only spaces and newlines are allowed; further, as we cannot
use the `libprocess` sub-processing facilities, we cannot timeout
the script, should this block for long times (or even forever).

This will conflict with the --ip flag, so if both set, we will fail
with an appropriate error message.

Review: https://reviews.apache.org/r/36425