Added a new API call 'updateAvailable' to the allocator.
Needed to implement the master HTTP endpoints: `/reserve`, `/unreserve`,
`/create` and `/destroy`.

This API is similar to `updateSlave` which is currently very specific to
oversubscription.  I considered consolidating `updateAvailable` and
`updateSlave` but it will require making offers be generated within the
allocator to enable this.

In specific, `updateAvailable` could fail if there aren't sufficient
available resources to update, whereas `updateSlave` avoids failing by
keeping the allocator in an over-allocated state. For `updateSlave`,
leaving the allocator in an over-allocated state is ok. This is because
it does not modify resources therefore `total - allocated` will work out
to __empty__. `updateAvailable` cannot leave the allocator in an
over-allocated state, because it modifies resources, and therefore
`total - allocated` is not guaranteed to yield __empty__.

Review: https://reviews.apache.org/r/35947