Updated operational guide to document standalone to HA transition.
This methodology has been discuessed in the mailing list
and I've tested this several times with a dev clsuter.

Most recently, I've used this procedure to migrate a production
mesos cluster (with over 100 services and crons, and 6 mesos slaves)
from _standalone_ to `--quorum=2` with no issues.  It is probably
worth noting that the only framework in use on the cluster is Aurora.

Review: https://reviews.apache.org/r/35361